another int => bigint issue in addressline

+++ sql/functions.sql
@@ -2254,7 +2254,7 @@ DROP TYPE addressline CASCADE;
 create type addressline as (
   place_id BIGINT,
   osm_type CHAR(1),
-  osm_id INTEGER,
+  osm_id BIGINT,
   name HSTORE,
   class TEXT,
   type TEXT,
